Fix a typo.
When prepending LLVM to the current path, an extra brace was causing the
wrong LLVM (and, specificially, the wrong `llvm-config`) to get picked
up by subsequent builds. This could lead to version mismatches, which
manifested themselves as dynamic linking errors when symbol names from
core LLVM libraries didn't quite match those referenced from Loom and
llvm-prov libraries.

Unfortunately this error might not be noticed when the path already
happened to contain the correct version of llvm-config, which is why it
slipped through.